+++ code/src/nuvla/job/job.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 
+import time
 import logging
 
 from nuvla.api import NuvlaError, ConnectionError
 
-from .util import wait, retry_kazoo_queue_op
+from .util import retry_kazoo_queue_op
 
 log = logging.getLogger('job')
 
@@ -33,18 +34,22 @@ class Job(dict):
 
     def _init(self):
         try:
-            self.id = self.queue.get().decode()
-            cimi_job = self.get_cimi_job(self.id)
-            dict.__init__(self, cimi_job)
-            if self.is_in_final_state():
-                retry_kazoo_queue_op(self.queue, "consume")
-                log.warning('Newly retrieved {} already in final state! Removed from queue.'
-                            .format(self.id))
+            self.id = self.queue.get(timeout=5)
+            if self.id is None:
                 self.nothing_to_do = True
-            elif self.get('state') == 'RUNNING':
-                # could happen when updating job and cimi server is down!
-                # let job actions decide what to do with it.
-                log.warning('Newly retrieved {} in running state!'.format(self.id))
+            else:
+                self.id = self.id.decode()
+                cimi_job = self.get_cimi_job(self.id)
+                dict.__init__(self, cimi_job)
+                if self.is_in_final_state():
+                    retry_kazoo_queue_op(self.queue, "consume")
+                    log.warning('Newly retrieved {} already in final state! Removed from queue.'
+                                .format(self.id))
+                    self.nothing_to_do = True
+                elif self.get('state') == 'RUNNING':
+                    # could happen when updating job and cimi server is down!
+                    # let job actions decide what to do with it.
+                    log.warning('Newly retrieved {} in running state!'.format(self.id))
         except NonexistentJobError as e:
             retry_kazoo_queue_op(self.queue, "consume")
             log.warning('Newly retrieved {} does not exist in cimi; '.format(self.id) +
@@ -57,7 +62,7 @@ class Job(dict):
                 'Fatal error when trying to retrieve {}! Put it back in queue. '.format(self.id) +
                 'Will go back to work after {}s.'.format(timeout))
             log.exception(e)
-            wait(timeout)
+            time.sleep(timeout)
             self.nothing_to_do = True
 
     def get_cimi_job(self, job_uri):
@@ -72,7 +77,7 @@ class Job(dict):
                 if e.response.status_code == 404:
                     log.warning('Retrieve of {} failed. Attempt: {} Will retry in {}s.'
                                 .format(job_uri, attempt, wait_time))
-                    wait(wait_time)
+                    time.sleep(wait_time)
                 else:
                     raise e
         raise NonexistentJobError(reason)
@@ -120,7 +125,7 @@ class Job(dict):
 
         if has_to_update:
             self._edit_job('affected-resources',
-                           [{'href': id} for id in current_affected_resources_ids])
+                           [{'href': res_id} for res_id in current_affected_resources_ids])
 
     def update_job(self, state=None, return_code=None, status_message=None):
         attributes = {}
@@ -160,8 +165,7 @@ class Job(dict):
         except (NuvlaError, ConnectionError):
             retry_kazoo_queue_op(self.queue, 'release')
             reason = 'Failed to update following attributes "{}" for {}! ' \
-                         .format(attributes, self.id) \
-                     + 'Put it back in queue.'
+                     'Put it back in queue.'.format(attributes, self.id)
             raise JobUpdateError(reason)
         else:
             self.update(response.data)